# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ice Cream Calculator API",
    "version": "1.0.0",
    "description": "API for managing ice cream ingredients and suppliers"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Allergens": {
        "type": "object",
        "properties": {
          "wheat": {
            "type": "boolean"
          },
          "fish": {
            "type": "boolean"
          },
          "crustacean": {
            "type": "boolean"
          },
          "mollusc": {
            "type": "boolean"
          },
          "egg": {
            "type": "boolean"
          },
          "milk": {
            "type": "boolean"
          },
          "lupin": {
            "type": "boolean"
          },
          "peanut": {
            "type": "boolean"
          },
          "soy": {
            "type": "boolean"
          },
          "sesame": {
            "type": "boolean"
          },
          "almond": {
            "type": "boolean"
          },
          "brazilNut": {
            "type": "boolean"
          },
          "cashew": {
            "type": "boolean"
          },
          "hazelnut": {
            "type": "boolean"
          },
          "macadamia": {
            "type": "boolean"
          },
          "pecan": {
            "type": "boolean"
          },
          "pistachio": {
            "type": "boolean"
          },
          "pineNut": {
            "type": "boolean"
          },
          "walnut": {
            "type": "boolean"
          },
          "barley": {
            "type": "boolean"
          },
          "oats": {
            "type": "boolean"
          },
          "rye": {
            "type": "boolean"
          },
          "sulphites": {
            "type": "boolean"
          }
        },
        "required": [
          "wheat",
          "fish",
          "crustacean",
          "mollusc",
          "egg",
          "milk",
          "lupin",
          "peanut",
          "soy",
          "sesame",
          "almond",
          "brazilNut",
          "cashew",
          "hazelnut",
          "macadamia",
          "pecan",
          "pistachio",
          "pineNut",
          "walnut",
          "barley",
          "oats",
          "rye",
          "sulphites"
        ]
      },
      "IngredientCategory": {
        "type": "string",
        "enum": [
          "Mix-in",
          "Dried Fruit",
          "Other",
          "Flavour",
          "Fruit",
          "Sugar",
          "Alcohol",
          "Vegetable",
          "Seed",
          "Dairy",
          "Chocolate",
          "Oil",
          "Stabilizer",
          "Grain"
        ]
      },
      "SugarsPer100g": {
        "type": "object",
        "properties": {
          "sucrose": {
            "type": "number"
          },
          "fructose": {
            "type": "number"
          },
          "glucose": {
            "type": "number"
          },
          "dextrose": {
            "type": "number"
          },
          "alcohol": {
            "type": "number"
          },
          "other": {
            "type": "number"
          }
        },
        "required": [
          "sucrose",
          "fructose",
          "glucose",
          "dextrose",
          "alcohol",
          "other"
        ]
      },
      "SupplierContactInfo": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "contactInfo": {
            "$ref": "#/components/schemas/SupplierContactInfo"
          },
          "website": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "id",
          "name",
          "contactInfo"
        ]
      },
      "Ingredient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "supplierID": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Discontinued"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/IngredientCategory"
          },
          "type": {
            "type": "string",
            "enum": [
              "Wet",
              "Dry"
            ]
          },
          "brand": {
            "type": "string"
          },
          "foodCompositionID": {
            "type": "string"
          },
          "energyPer100g": {
            "type": "number",
            "minimum": 0
          },
          "proteinPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "saturatedFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalCarbPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalSugarsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "sodiumMgPer100g": {
            "type": "number",
            "minimum": 0
          },
          "waterPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "sugarsPer100g": {
            "$ref": "#/components/schemas/SugarsPer100g"
          },
          "totalSolidsPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "otherSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "MSNFPer100g": {
            "type": "number",
            "minimum": 0
          },
          "PAC": {
            "type": "number",
            "minimum": 0
          },
          "POD": {
            "type": "number",
            "minimum": 0
          },
          "HF": {
            "type": "number",
            "minimum": 0
          },
          "dryCocoaSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "cocoaButterPer100g": {
            "type": "number",
            "minimum": 0
          },
          "supplier": {
            "$ref": "#/components/schemas/Supplier"
          },
          "supplierCode": {
            "type": "string"
          },
          "packageSizeInGrams": {
            "type": "number",
            "minimum": 1
          },
          "costPerPackInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "costPer1000gInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "percentOfUsefulProduct": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "allergens": {
            "$ref": "#/components/schemas/Allergens"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "supplierID",
          "status",
          "category",
          "type",
          "brand",
          "foodCompositionID",
          "energyPer100g",
          "proteinPer100g",
          "totalFatPer100g",
          "saturatedFatPer100g",
          "totalCarbPer100g",
          "totalSugarsPer100g",
          "sodiumMgPer100g",
          "waterPer100g",
          "sugarsPer100g",
          "totalSolidsPer100g",
          "otherSolidsPer100g",
          "MSNFPer100g",
          "PAC",
          "POD",
          "HF",
          "dryCocoaSolidsPer100g",
          "cocoaButterPer100g",
          "supplier",
          "supplierCode",
          "packageSizeInGrams",
          "costPerPackInCentsExGST",
          "costPer1000gInCentsExGST",
          "percentOfUsefulProduct",
          "allergens",
          "createdAt",
          "lastModifiedAt"
        ]
      },
      "CreateIngredientRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "supplierID": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Discontinued"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/IngredientCategory"
          },
          "type": {
            "type": "string",
            "enum": [
              "Wet",
              "Dry"
            ]
          },
          "brand": {
            "type": "string",
            "minLength": 1
          },
          "foodCompositionID": {
            "type": "string",
            "minLength": 1
          },
          "energyPer100g": {
            "type": "number",
            "minimum": 0
          },
          "proteinPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "saturatedFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalCarbPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalSugarsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "sodiumMgPer100g": {
            "type": "number",
            "minimum": 0
          },
          "waterPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "sugarsPer100g": {
            "$ref": "#/components/schemas/SugarsPer100g"
          },
          "totalSolidsPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "otherSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "MSNFPer100g": {
            "type": "number",
            "minimum": 0
          },
          "PAC": {
            "type": "number",
            "minimum": 0
          },
          "POD": {
            "type": "number",
            "minimum": 0
          },
          "HF": {
            "type": "number",
            "minimum": 0
          },
          "dryCocoaSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "cocoaButterPer100g": {
            "type": "number",
            "minimum": 0
          },
          "supplierCode": {
            "type": "string",
            "minLength": 1
          },
          "packageSizeInGrams": {
            "type": "number",
            "minimum": 1
          },
          "costPerPackInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "costPer1000gInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "percentOfUsefulProduct": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "allergens": {
            "$ref": "#/components/schemas/Allergens"
          }
        },
        "required": [
          "name",
          "supplierID",
          "status",
          "category",
          "type",
          "brand",
          "foodCompositionID",
          "energyPer100g",
          "proteinPer100g",
          "totalFatPer100g",
          "saturatedFatPer100g",
          "totalCarbPer100g",
          "totalSugarsPer100g",
          "sodiumMgPer100g",
          "waterPer100g",
          "sugarsPer100g",
          "totalSolidsPer100g",
          "otherSolidsPer100g",
          "MSNFPer100g",
          "PAC",
          "POD",
          "HF",
          "dryCocoaSolidsPer100g",
          "cocoaButterPer100g",
          "supplierCode",
          "packageSizeInGrams",
          "costPerPackInCentsExGST",
          "costPer1000gInCentsExGST",
          "percentOfUsefulProduct",
          "allergens"
        ]
      },
      "UpdateIngredientRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "supplierID": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Discontinued"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/IngredientCategory"
          },
          "type": {
            "type": "string",
            "enum": [
              "Wet",
              "Dry"
            ]
          },
          "brand": {
            "type": "string",
            "minLength": 1
          },
          "foodCompositionID": {
            "type": "string",
            "minLength": 1
          },
          "energyPer100g": {
            "type": "number",
            "minimum": 0
          },
          "proteinPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "saturatedFatPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalCarbPer100g": {
            "type": "number",
            "minimum": 0
          },
          "totalSugarsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "sodiumMgPer100g": {
            "type": "number",
            "minimum": 0
          },
          "waterPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "sugarsPer100g": {
            "$ref": "#/components/schemas/SugarsPer100g"
          },
          "totalSolidsPer100g": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "otherSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "MSNFPer100g": {
            "type": "number",
            "minimum": 0
          },
          "PAC": {
            "type": "number",
            "minimum": 0
          },
          "POD": {
            "type": "number",
            "minimum": 0
          },
          "HF": {
            "type": "number",
            "minimum": 0
          },
          "dryCocoaSolidsPer100g": {
            "type": "number",
            "minimum": 0
          },
          "cocoaButterPer100g": {
            "type": "number",
            "minimum": 0
          },
          "supplierCode": {
            "type": "string",
            "minLength": 1
          },
          "packageSizeInGrams": {
            "type": "number",
            "minimum": 1
          },
          "costPerPackInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "costPer1000gInCentsExGST": {
            "type": "number",
            "minimum": 0
          },
          "percentOfUsefulProduct": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "allergens": {
            "$ref": "#/components/schemas/Allergens"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "number",
            "minimum": 1
          },
          "totalPages": {
            "type": "number",
            "minimum": 0
          },
          "totalCount": {
            "type": "number",
            "minimum": 0
          },
          "pageSize": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "currentPage",
          "totalPages",
          "totalCount",
          "pageSize"
        ]
      },
      "GetIngredientsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "ingredients": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/PaginationMeta"
              }
            },
            "required": [
              "ingredients",
              "pagination"
            ]
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "CreateIngredientResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Ingredient"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "GetSuppliersResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Supplier"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "success",
          "error"
        ]
      }
    }
  },
  "paths": {
    "/ingredients": {
      "get": {
        "summary": "Get all ingredients",
        "description": "Retrieve a paginated list of all ingredients with optional filtering",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for ingredient name or category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by ingredient status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive",
                "Discontinued"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by ingredient category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by ingredient type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Wet",
                "Dry"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIngredientsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new ingredient",
        "description": "Create a new ingredient with full nutritional and supplier information",
        "tags": [
          "Ingredients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ingredient created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIngredientResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Ingredient already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingredients/{id}": {
      "get": {
        "summary": "Get ingredient by ID",
        "description": "Retrieve a specific ingredient by its ID",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Ingredient ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Ingredient"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Ingredient not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update ingredient",
        "description": "Update an existing ingredient",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Ingredient ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIngredientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingredient updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Ingredient"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ingredient not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete ingredient",
        "description": "Delete an existing ingredient",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Ingredient ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredient deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Ingredient not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/suppliers": {
      "get": {
        "summary": "Get all suppliers",
        "description": "Retrieve a list of all suppliers",
        "tags": [
          "Suppliers"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSuppliersResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}